// generated/src/app/api-models/service-area.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ServiceArea {
  serviceAreaName: string;
  serviceAreaDescription: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}